Refactoring DateBubble component so that date range is only displayed when end date prop is specified
