+++ src/shared/util/events.js
@@ -4,10 +4,8 @@ export function splitEvents({
   events,              // array of events
   timeline,            // one of 'past', 'today', 'future'
   reverse = false,     // optionally reverse final list of events
-  todayDateTime = null,  // iso string representing today date
+  todayDateTime = new Date(),  // iso string representing today date
 }) {
-  const today = todayDateTime || new Date();
-
   let relevantEvents = events.filter(event => {
     const startDateTime = dateFns.parse(event.startDateTime.iso);
     const endDateTime = dateFns.parse(event.endDateTime.iso);
@@ -21,15 +19,15 @@ export function splitEvents({
     switch (timeline) {
       case 'today':
         if (dateFns.isSameDay(startDateTime, endDateTime)) {
-          return dateFns.isSameDay(startDateTime, today);
+          return dateFns.isSameDay(startDateTime, todayDateTime);
         } else { // eslint-disable-line no-else-return
-          return dateFns.isWithinRange(today, startDateTime, endDateTime);
+          return dateFns.isWithinRange(todayDateTime, startDateTime, endDateTime);
         }
       case 'past':
-        return (!dateFns.isWithinRange(today, startDateTime, endDateTime) && dateFns.isBefore(endDateTime, today) && !dateFns.isSameDay(endDateTime, today));
+        return (!dateFns.isWithinRange(todayDateTime, startDateTime, endDateTime) && dateFns.isBefore(endDateTime, todayDateTime) && !dateFns.isSameDay(endDateTime, todayDateTime));
       case 'future':
-        return (!dateFns.isSameDay(startDateTime, today) &&
-          dateFns.isAfter(startDateTime, today));
+        return (!dateFns.isSameDay(startDateTime, todayDateTime) &&
+          dateFns.isAfter(startDateTime, todayDateTime));
       default:
         return false;
     }
@@ -40,4 +38,15 @@ export function splitEvents({
   }
 
   return relevantEvents;
+}
+
+// Helper function for displaying date range for multi day events
+// Date range would only be displayed if end date is set on the
+// child component
+export function setEndDate(timeline, startDateTime, endDateTime) {
+  if (timeline === 'today' &&
+    startDateTime.date !== endDateTime.date) {
+    return endDateTime;
+  }
+  return null;
 }
